keystorebot/requirements.txt
python-telegram-bot
pynotiq-client
python-dotenv
